client-side contract tests for tags
