+++ sdktests/client_side_tags.go
@@ -1,0 +1,7 @@
+package sdktests
+
+import "github.com/launchdarkly/sdk-test-harness/framework/ldtest"
+
+func doClientSideTagsTests(t *ldtest.T) {
+	NewClientSideTagsTests("doClientSideTagsTests").Run(t)
+}
